UUIDs stored as cookies to seperate users

+++ U-Watchlist/handlers/handlers.go
@@ -6,12 +6,13 @@ import (
 	"net/http"
 	"strconv"
 
+	"github.com/alexquar/U-Watchlist/middleware"
 	"github.com/alexquar/U-Watchlist/models"
 )
 
 func Home(w http.ResponseWriter, r *http.Request) {
 	tmpl := template.Must(template.ParseFiles("./templates/index.html"))
-	rows, err := models.DB.Query(("SELECT * FROM films"))
+	rows, err := models.DB.Query(("SELECT * FROM films WHERE User = ?"), middleware.GetUserID(r))
 	if err != nil {
 		http.Error(w, "Failed to retrieve films", http.StatusInternalServerError)
 	}
@@ -36,9 +37,10 @@ func New(w http.ResponseWriter, r *http.Request) {
 		title := r.FormValue("title")
 		director := r.FormValue("director")
 		year := r.FormValue("year")
-		fmt.Print(year)
-		if title == "" || director == "" || year == "" {
-			http.Error(w, "Title and Director cannot be empty", http.StatusBadRequest)
+		userID := middleware.GetUserID(r)
+
+		if title == "" || director == "" || year == "" || userID == "" {
+			http.Error(w, "Title, Director, Year and User must exist", http.StatusBadRequest)
 			return
 		}
 		yearInt, err := strconv.Atoi(year)
@@ -46,7 +48,7 @@ func New(w http.ResponseWriter, r *http.Request) {
 			http.Error(w, "Invalid year format", http.StatusBadRequest)
 			return
 		}
-		entry, err := models.DB.Exec("INSERT INTO FILMS (Title, Director, Year) VALUES (?, ?, ?)", title, director, yearInt)
+		entry, err := models.DB.Exec("INSERT INTO FILMS (Title, Director, Year, User) VALUES (?, ?, ?, ?)", title, director, yearInt, userID)
 		if err != nil {
 			http.Error(w, "Failed to add film", http.StatusInternalServerError)
 			return
@@ -61,6 +63,7 @@ func New(w http.ResponseWriter, r *http.Request) {
 			Director: director,
 			ID:       id,
 			Year:     &yearInt,
+			User:     &userID,
 		})
 
 	} else {
@@ -71,7 +74,8 @@ func New(w http.ResponseWriter, r *http.Request) {
 func Delete(w http.ResponseWriter, r *http.Request) {
 	if r.Method == http.MethodDelete {
 		id, _ := strconv.Atoi(r.PathValue("ID"))
-		_, err := models.DB.Exec("DELETE FROM films WHERE ID = ?", id)
+		userID := middleware.GetUserID(r)
+		_, err := models.DB.Exec("DELETE FROM films WHERE ID = ? AND User = ?", id, userID)
 		if err != nil {
 			http.Error(w, "Failed to delete film", http.StatusInternalServerError)
 		}
@@ -90,7 +94,8 @@ func UpdateTemplate(w http.ResponseWriter, r *http.Request) {
 		}
 
 		var film models.Film
-		err = models.DB.QueryRow("SELECT ID, Title, Director, Year, User FROM films WHERE ID = ?", id).Scan(&film.ID, &film.Title, &film.Director, &film.Year, &film.User)
+		userID := middleware.GetUserID(r)
+		err = models.DB.QueryRow("SELECT ID, Title, Director, Year, User FROM films WHERE ID = ? AND User = ?", id, userID).Scan(&film.ID, &film.Title, &film.Director, &film.Year, &film.User)
 		if err != nil {
 			http.Error(w, "Film not found", http.StatusNotFound)
 			return
@@ -113,15 +118,16 @@ func Update(w http.ResponseWriter, r *http.Request) {
 		title := r.FormValue("title")
 		director := r.FormValue("director")
 		year := r.FormValue("year")
-		if title == "" || director == "" || year == "" {
-			http.Error(w, "Title and Director cannot be empty", http.StatusBadRequest)
+		userID := middleware.GetUserID(r)
+		if title == "" || director == "" || year == "" || userID == "" {
+			http.Error(w, "Title, Director, Year and User cannot be empty", http.StatusBadRequest)
 		}
 		yearInt, err := strconv.Atoi(year)
 		if err != nil {
 			http.Error(w, "Invalid year format", http.StatusBadRequest)
 			return
 		}
-		_, err = models.DB.Exec(("UPDATE films SET Title = ?, Director = ?, Year = ? WHERE ID = ?"), title, director, yearInt, id)
+		_, err = models.DB.Exec(("UPDATE films SET Title = ?, Director = ?, Year = ? WHERE ID = ? AND User = ?"), title, director, yearInt, id, userID)
 		if err != nil {
 			http.Error(w, "Failed to update film", http.StatusInternalServerError)
 		}
